Rolled all the constructors into one file
Need to rethink this though, we have to have imports in functions right
now because of circular imports
